Changes to Bucket Manager to handle multiple resources.

+++ Master40.SimulationImmutables/Message/FSetupDefinition.fs
@@ -3,6 +3,6 @@
 open Akka.Actor
 
 type FSetupDefinition = {
-    SetupKey : int64
+    SetupKey : int32
     RequiredResources : System.Collections.Generic.List<IActorRef>
 }
